Automatic client address learning:
When we receive a good packet from a client, extract the senders TUN IP from the underlying IP packet. This is stored in a HashMap as a key and the public facing IP address(es) are stored as a Vec<sockaddr> value.

Now, how do we use this when responding?