Automatic client address learning:
The configuration file now supports a "remote_tun_addr" entry. If it is present, it will be used to seed the client_list.
Whenever a packet is to be sent, the sender will lookup the destination tunnel port in the client_list. If it exists, packet data will be sent to all existing target addresses in the list.