Connection keepalive
In situations where the clients are communicating through a NAT we need to keep the PAT (Port Address Translation) open. This is can hopefully be done by introducing a periodic keepalive.

The keepalive is controlled by two options in the configuration JSON:
 * keep_alive: Boolean
 * keep_alive_interval: u64
If these options are not available or keep_alive is set to false, then keepalive will not be used.